ai_supermarket/requirements.txt
streamlit
pandas
pyyaml
ibm-watsonx-orchestrate>=1.6.0
requests
python-dotenv
flask
flask-cors